refactor: answers의 delete가 검증과 삭제의 책임을 모두 갖도록 수정

+++ src/main/java/nextstep/qna/domain/Answers.java
@@ -21,18 +21,18 @@ public class Answers {
         answers.add(answer);
     }
 
-    public void checkDeletable(NsUser loginUser) throws CannotDeleteException {
-        if (answers.stream().anyMatch(answer -> !answer.isOwner(loginUser))) {
-            throw new CannotDeleteException("다른 사람의 답변이 존재하여 삭제할 수 없습니다.");
-        }
-    }
-
-    public List<DeleteHistory> delete() {
+    public List<DeleteHistory> delete(NsUser loginUser) throws CannotDeleteException {
+        checkDeletable(loginUser);
         List<DeleteHistory> histories = new ArrayList<>();
         for (Answer answer : answers) {
             histories.add(answer.delete());
         }
         return histories;
+    }
+    private void checkDeletable(NsUser loginUser) throws CannotDeleteException {
+        if (answers.stream().anyMatch(answer -> !answer.isOwner(loginUser))) {
+            throw new CannotDeleteException("다른 사람의 답변이 존재하여 삭제할 수 없습니다.");
+        }
     }
 
     public List<Answer> getAnswers() {
